+++ CHANGELOG.md
@@ -1,5 +1,10 @@
+# Unreleased
+
+Added
+  * AZ-770: Add Terraform module info in output
+
 # v5.2.0 - 2022-06-10
- 
+
 Added
   * AZ-775: Added `force_ssl_minimal_version` variable to fix plan with force_ssl=false