+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Unreleased
+# v6.2.1 - 2023-03-01
 
 Fixed
   * AZ-989: Add `pre-commit` and update CI